# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs by Manuel Almaguer",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "Login User Here, default user: <b>welove@tracktik.com</b>,  password: <b>welovetracktik</b> ",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/question/one": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Question one",
                "description": "Question one here",
                "operationId": "Scenario",
                "responses": {
                    "200": {
                        "description": "Success Scenario one",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/question/two?type=console": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Question two",
                "description": "Question two here",
                "operationId": "ScenarioTwo",
                "responses": {
                    "200": {
                        "description": "Success Scenario two",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/types": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "Return types",
                "description": "Return types",
                "operationId": "Types",
                "responses": {
                    "200": {
                        "description": "Return types",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "PASTE THIS INTO THE INPUT BELOW => <b>Bearer dummytokenjustforexample</b>  <br /> If you change de token from the .env, please paste usgin de previus format",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}